notes
considering the cranfield folder, we will run each of the queries in the queries file on the search engine, and from the results, decide what the AP is of each query.
The lines in qrel refer to the document name (not ID). and also correspond to each query in each line as the queries file.
We will build code to read all of the lines of both files into a list, run query #1, get the results back from the search engine and use the qrel to determine if each of the results are correct or not to compute AP.
We will be doing this same thing on the parks folder.
The relevant method I made passes in the index (as an integer), query (as a list), and qrel (as a list).
The index corresponds to each index in both lists to check whether the query at index i is relevant by checking if the document name is in qrel at index i.